[Refactor] 패키지 수정
- common 패키지 생성
- common안에 utils,base,network
- extension 삭제
- utils 안에 ext 파일 넣기
