+++ app/src/main/java/com/doctor/yumyum/common/base/activity/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.doctor.yumyum.presentation.base.activity
+package com.doctor.yumyum.common.base.activity
 
 import android.os.Bundle
 import android.widget.Toast
