--- app/src/main/java/com/doctor/yumyum/extension/TimeExtension.kt
@@ -1,1 +1,0 @@
-package com.doctor.yumyum.extension
